Replace cp command with uutils cp version
Rename current cp to CpOld
