+++ crates/nu-command/Cargo.toml
@@ -89,9 +89,8 @@ unicode-segmentation = "1.10"
 ureq = { version = "2.7", default-features = false, features = ["charset", "gzip", "json", "native-tls"] }
 url = "2.2"
 # uu_cp = { path = "../../../coreutils/src/uu/cp"}
+# 0.0.21 not on crates.io yet?
 uu_cp = { git = "https://github.com/uutils/coreutils.git", rev = "862a2df"}
-# uucore = { git = "https://github.com/uutils/coreutils.git", rev = "1f081f3"}
-# uucore = { path = "../../../coreutils/src/uucore"}
 uuid = { version = "1.3", features = ["v4"] }
 wax = { version = "0.5" }
 which = { version = "4.4", optional = true }
@@ -121,7 +120,6 @@ features = [
 version = "0.48"
 
 [features]
-nuuu = []
 plugin = ["nu-parser/plugin"]
 sqlite = ["rusqlite"]
 trash-support = ["trash"]
