+++ crates/nu-command/src/default_context.rs
@@ -196,17 +196,6 @@ pub fn add_shell_command_context(mut engine_state: EngineState) -> EngineState {
             FormatDuration,
             FormatFilesize,
         };
-        // UUtils Support
-        #[cfg(feature = "nuuu")]
-        bind_command! {
-            Ucp,
-        }
-
-        // FileSystem non-nuuu
-        #[cfg(not(feature = "nuuu"))]
-        bind_command! {
-            Cp,
-        };
 
         // FileSystem
         bind_command! {
@@ -214,6 +203,8 @@ pub fn add_shell_command_context(mut engine_state: EngineState) -> EngineState {
             Ls,
             Mkdir,
             Mv,
+            Cp,
+            CpOld,
             Open,
             Start,
             Rm,
