+++ crates/nu-command/src/filesystem/mod.rs
@@ -1,6 +1,7 @@
 mod cd;
 mod cd_query;
 mod cp;
+mod cp_old;
 mod glob;
 mod ls;
 mod mkdir;
@@ -10,8 +11,6 @@ mod rm;
 mod save;
 mod start;
 mod touch;
-#[cfg(feature = "nuuu")]
-mod ucp;
 mod util;
 mod watch;
 
@@ -19,6 +18,7 @@ pub use self::open::Open;
 pub use cd::Cd;
 pub use cd_query::query;
 pub use cp::Cp;
+pub use cp_old::CpOld;
 pub use glob::Glob;
 pub use ls::Ls;
 pub use mkdir::Mkdir;
@@ -27,6 +27,4 @@ pub use rm::Rm;
 pub use save::Save;
 pub use start::Start;
 pub use touch::Touch;
-#[cfg(feature = "nuuu")]
-pub use ucp::Ucp;
 pub use watch::Watch;
